Fixed small compiling warning
