+++ utils.h
@@ -3,6 +3,9 @@
 
 #include <string>
 
+#include <sys/stat.h>
+#include <sys/statvfs.h>
+
 #include "logger/Logger.h"
 
 namespace os {
@@ -19,8 +22,20 @@ namespace os {
 	};
 
 	void check_or_create(const string& path, Logger* logger = NULL);
-	inline bool file_exists(const string& name);
-	inline float get_available_disk_space();
+
+	inline bool file_exists(const string& name)
+	{
+		struct stat buffer;
+		return stat(name.c_str(), &buffer) == 0;
+	}
+
+	inline float get_available_disk_space()
+	{
+		struct statvfs fs;
+		statvfs("data", &fs);
+
+		return ((float) fs.f_bsize)*fs.f_bavail;
+	}
 
 	State set_state(State new_state);
 	State get_last_state();
